scripts/hierarchical/requirements.txt
pandas
plotly.express
scikit-learn
azure.storage.blob
pyarrow
setuptools